Fix missing token information on minting (registers were empty)

+++ FleetSharp/Sigma/DataSerializer.cs
@@ -118,8 +118,8 @@ namespace FleetSharp.Sigma
             }
             else if (Utils.isColl(data))
             {
-                if (((ISigmaCollection)data).value.GetType() == typeof(string)) writer.writeVlq(((ISigmaCollection)data).value.Length / 2);
-                else writer.writeVlq(((ISigmaCollection)data).value.Length);
+                if (((ISigmaCollection)data).value.GetType() == typeof(string)) writer.writeVlq((uint)(((ISigmaCollection)data).value.Length / 2));
+                else writer.writeVlq((uint)((ISigmaCollection)data).value.Length);
 
                 switch (((ISigmaCollection)data).elementsType)
                 {
